Better child cleanup
This allows reflex to handle commands which run as root. For
instance:

reflex -s -- sh -c 'make && sudo ./myserver'

Reflex could not, before, kill the spawned child and grandchild
running as root, but now it hooks the job into a pty and can kill it
by sending a terminal control character (directly mimicing ^C).

We can no longer distinguish stderr/stdout (not sure how to make that
work with this new pty scheme) but I don't think it's actually
important.